requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
pydantic==2.12.5
alembic==1.18.4
factory_boy==3.3.3
pytest==9.0.2
pytest-asyncio==1.3.0

black==24.8.0
flake8==4.0.1
flake8-bugbear==22.1.11
flake8-pie==0.15.0
isort==5.9.3
mypy==0.931